+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Changed
+- collection published by date, updated filter and ordering
 
 ## [2.4.1] - 2020-04-09
+
 ### Fixed
 - add panorama image dir to deployer config